==> cardiac_gwas_coloc/1.1.process_sumstats.ipynb <==
# Process summary statistics
- decompress and compress again (bgz) the summary statistics files
- tabix the files


setwd("/frazer01/projects/CARDIPS/analysis/cardiac_gwas_coloc")

source("script/functions.R"  )


dir.create("pipeline/1.1.sumstats"         , showWarnings = FALSE)
dir.create("pipeline/1.1.sumstats/sumstats", showWarnings = FALSE)


infiles = list.files("input/sumstats")
infiles = data.frame(gz  = paste(getwd(), "input"   ,                 "sumstats", infiles                       , sep = "/"), 
                     bgz = paste(getwd(), "pipeline", "1.1.sumstats", "sumstats", sub("gz$", "bgz"    , infiles), sep = "/"), 
                     tbi = paste(getwd(), "pipeline", "1.1.sumstats", "sumstats", sub("gz$", "bgz.tbi", infiles), sep = "/"))



tabix_sumstats = function(ii, infiles)
{
    gz  = infiles[ii, "gz" ]
    bgz = infiles[ii, "bgz"]
    tbi = infiles[ii, "tbi"]
    
    message(gz)
    
    command_bgz = paste("gzip" , "-d", "--to-stdout", gz, "|", "bgzip", "-@12", ">", bgz)
    command_tbi = paste("tabix", "-b", 2, "-e", 2, "-s", 1, "-S", 1, bgz)
    
    system(command_bgz)
    system(command_tbi)
}

ii = 1
invisible(lapply(1:nrow(infiles), function(ii){tabix_sumstats(ii, infiles)}))
